complete tick (sub-frame) support
